requirements.txt
# core libraries
openai
anthropic
openai-agents[litellm]
langchain
pinecone
gradio
python-dotenv
